# artifacts/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Grid-Guard NTL Detection API",
    "description": "Production-grade REST API for smart-meter non-technical loss (NTL) detection, cost-sensitive economic decisioning, and SHAP explainability.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/predict": {
      "post": {
        "tags": [
          "Inference & Prediction"
        ],
        "summary": "Evaluate Single Smart Meter",
        "description": "Ingests a raw daily smart-meter time series, calculates 60 canonical temporal features, scores the meter with the cost-sensitive LightGBM champion booster, evaluates financial exposure and Expected Net Value (ENV), and optionally computes Tree-SHAP local attributions and domain tampering narratives.",
        "operationId": "predict_single_meter_api_v1_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SingleMeterPredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful inference and decision response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleMeterPredictionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Insufficient history or malformed input readings"
          },
          "422": {
            "description": "Pydantic validation failure"
          },
          "503": {
            "description": "Model service unavailable"
          }
        }
      }
    },
    "/api/v1/predict/batch": {
      "post": {
        "tags": [
          "Inference & Prediction"
        ],
        "summary": "Evaluate Batch of Smart Meters",
        "description": "Synchronously processes a bounded batch of smart-meter requests (up to configured maximum, e.g. 50 meters). Supports partial success: valid meters return predictions while malformed meters return structured per-item error records.",
        "operationId": "predict_batch_meters_api_v1_predict_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchPredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Batch processing completed with item-level results and errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchPredictionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Batch size exceeds configured maximum"
          },
          "422": {
            "description": "Batch schema validation failure"
          },
          "503": {
            "description": "Model service unavailable"
          }
        }
      }
    },
    "/api/v1/inspection/ticket": {
      "post": {
        "tags": [
          "Inspection & Field Dispatch"
        ],
        "summary": "Generate Field Inspection Work Order Ticket",
        "description": "Evaluates the target meter and compiles a complete, actionable field inspection ticket. Includes deterministic ticket ID, probability risk, estimated recoverable revenue, dispatch cost, Expected Net Value (ENV), active threshold, detected electrical tampering signatures, and audited human-readable narratives.",
        "operationId": "generate_inspection_ticket_api_v1_inspection_ticket_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SingleMeterPredictionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Inspection ticket generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionTicketResponse"
                }
              }
            }
          },
          "400": {
            "description": "Insufficient history or invalid meter payload"
          },
          "422": {
            "description": "Pydantic validation failure"
          },
          "503": {
            "description": "Model service unavailable"
          }
        }
      }
    },
    "/api/v1/inspection/queue": {
      "post": {
        "tags": [
          "Inspection & Field Dispatch"
        ],
        "summary": "Generate Prioritized Field Inspection Queue",
        "description": "Retrieves and filters candidate inspection tickets ranked strictly by Expected Net Value (ENV) descending. Allows dynamic filtering by minimum ENV threshold, minimum probability cutoff, and maximum crew capacity limit.",
        "operationId": "get_inspection_queue_api_v1_inspection_queue_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectionQueueRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ranked inspection queue returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectionQueueResponse"
                }
              }
            }
          },
          "404": {
            "description": "Precomputed evaluation artifacts not found"
          },
          "503": {
            "description": "Model service unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BatchPredictionRequest": {
        "properties": {
          "meters": {
            "items": {
              "$ref": "#/components/schemas/SingleMeterPredictionRequest"
            },
            "type": "array",
            "minItems": 1,
            "title": "Meters",
            "description": "List of single meter prediction requests"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id",
            "description": "Optional client-provided batch correlation identifier"
          },
          "include_explanation": {
            "type": "boolean",
            "title": "Include Explanation",
            "description": "Default to False for batch to maximize throughput",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "meters"
        ],
        "title": "BatchPredictionRequest",
        "description": "Request payload for bounded multi-meter batch evaluation."
      },
      "BatchPredictionResponse": {
        "properties": {
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "total_items": {
            "type": "integer",
            "title": "Total Items"
          },
          "successful_items": {
            "type": "integer",
            "title": "Successful Items"
          },
          "failed_items": {
            "type": "integer",
            "title": "Failed Items"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SingleMeterPredictionResponse"
            },
            "type": "array",
            "title": "Results"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/BatchItemError"
            },
            "type": "array",
            "title": "Errors"
          },
          "processing_time_ms": {
            "type": "number",
            "title": "Processing Time Ms"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "total_items",
          "successful_items",
          "failed_items",
          "processing_time_ms"
        ],
        "title": "BatchPredictionResponse",
        "description": "Response payload for multi-meter batch requests supporting partial success."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InspectionQueueRequest": {
        "properties": {
          "max_inspections": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Max Inspections",
            "description": "Maximum number of prioritized tickets to return",
            "default": 50
          },
          "decision_rule": {
            "type": "string",
            "title": "Decision Rule",
            "description": "Decision policy: 'env', 'cost_threshold', or 'fixed_threshold'",
            "default": "env"
          },
          "min_env": {
            "type": "number",
            "title": "Min Env",
            "description": "Minimum Expected Net Value cutoff",
            "default": 0.0
          },
          "min_probability": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Probability",
            "description": "Minimum probability cutoff",
            "default": 0.5
          },
          "include_explanations": {
            "type": "boolean",
            "title": "Include Explanations",
            "description": "Whether to include full SHAP narratives for queued tickets",
            "default": true
          }
        },
        "type": "object",
        "title": "InspectionQueueRequest",
        "description": "Request parameters for querying or ranking inspection candidates into a prioritized queue."
      },
      "InspectionQueueResponse": {
        "properties": {
          "queue_size": {
            "type": "integer",
            "title": "Queue Size",
            "description": "Number of tickets in the returned inspection queue"
          },
          "total_expected_recovery": {
            "type": "number",
            "title": "Total Expected Recovery",
            "description": "Sum of expected gross recovery across all tickets ($)"
          },
          "total_dispatch_cost": {
            "type": "number",
            "title": "Total Dispatch Cost",
            "description": "Total crew dispatch cost for the queue ($)"
          },
          "total_net_value": {
            "type": "number",
            "title": "Total Net Value",
            "description": "Cumulative Expected Net Value (total recovery - total cost) ($)"
          },
          "decision_rule": {
            "type": "string",
            "title": "Decision Rule",
            "description": "Decision policy applied"
          },
          "tickets": {
            "items": {
              "$ref": "#/components/schemas/InspectionTicketResponse"
            },
            "type": "array",
            "title": "Tickets",
            "description": "Ranked inspection work orders"
          },
          "processing_time_ms": {
            "type": "number",
            "title": "Processing Time Ms",
            "description": "Processing time in milliseconds"
          }
        },
        "type": "object",
        "required": [
          "queue_size",
          "total_expected_recovery",
          "total_dispatch_cost",
          "total_net_value",
          "decision_rule",
          "processing_time_ms"
        ],
        "title": "InspectionQueueResponse",
        "description": "Response payload containing a ranked inspection queue."
      },
      "InspectionTicketResponse": {
        "properties": {
          "ticket_id": {
            "type": "string",
            "title": "Ticket Id",
            "description": "Deterministic ticket identifier (e.g. 'TCK-2016-10-30-EF550F26')"
          },
          "meter_id": {
            "type": "string",
            "title": "Meter Id",
            "description": "Target smart meter account identifier"
          },
          "evaluation_period": {
            "type": "string",
            "title": "Evaluation Period",
            "description": "Evaluation snapshot period"
          },
          "tamper_probability": {
            "type": "number",
            "title": "Tamper Probability",
            "description": "Predicted uncalibrated tampering probability"
          },
          "calibrated_probability": {
            "type": "number",
            "title": "Calibrated Probability",
            "description": "Calibrated posterior probability p_i"
          },
          "estimated_leakage_kwh": {
            "type": "number",
            "title": "Estimated Leakage Kwh",
            "description": "Estimated volume of unmetered electricity in kWh"
          },
          "estimated_recoverable_revenue": {
            "type": "number",
            "title": "Estimated Recoverable Revenue",
            "description": "Projected recoverable revenue R ($)"
          },
          "dispatch_cost": {
            "type": "number",
            "title": "Dispatch Cost",
            "description": "Field inspection dispatch cost C_FP ($)"
          },
          "expected_gross_recovery": {
            "type": "number",
            "title": "Expected Gross Recovery",
            "description": "Expected gross recovery E[R] = p * R ($)"
          },
          "env": {
            "type": "number",
            "title": "Env",
            "description": "Expected Net Value (ENV) = p * R - C_dispatch ($)"
          },
          "tau_cost": {
            "type": "number",
            "title": "Tau Cost",
            "description": "Bayes cost-optimal threshold C_dispatch / (C_dispatch + C_FN)"
          },
          "tau_env": {
            "type": "number",
            "title": "Tau Env",
            "description": "Economic breakeven threshold C_dispatch / R"
          },
          "active_threshold": {
            "type": "number",
            "title": "Active Threshold",
            "description": "Active probability threshold applied"
          },
          "decision_rule": {
            "type": "string",
            "title": "Decision Rule",
            "description": "Decision policy applied ('env', 'cost_threshold', 'fixed')"
          },
          "inspection_recommended": {
            "type": "boolean",
            "title": "Inspection Recommended",
            "description": "Whether inspection is economically justified"
          },
          "priority_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Rank",
            "description": "Queue rank ordered by ENV descending"
          },
          "customer_type": {
            "type": "string",
            "title": "Customer Type",
            "description": "Customer classification",
            "default": "standard"
          },
          "feeder_id": {
            "type": "string",
            "title": "Feeder Id",
            "description": "Feeder network identifier",
            "default": "unknown"
          },
          "data_quality_status": {
            "type": "string",
            "title": "Data Quality Status",
            "description": "Quality assessment",
            "default": "good"
          },
          "signatures_summary": {
            "type": "string",
            "title": "Signatures Summary",
            "description": "Summary of electrical tampering signatures",
            "default": "Multivariate Feature Anomaly"
          },
          "explanation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExplanationOutput"
              },
              {
                "type": "null"
              }
            ],
            "description": "Detailed explainability record"
          },
          "safety_caveat": {
            "type": "string",
            "title": "Safety Caveat",
            "description": "Standard legal/regulatory disclaimer",
            "default": "Model evidence indicates an anomalous consumption pattern and requires physical on-site inspection."
          }
        },
        "type": "object",
        "required": [
          "ticket_id",
          "meter_id",
          "evaluation_period",
          "tamper_probability",
          "calibrated_probability",
          "estimated_leakage_kwh",
          "estimated_recoverable_revenue",
          "dispatch_cost",
          "expected_gross_recovery",
          "env",
          "tau_cost",
          "tau_env",
          "active_threshold",
          "decision_rule",
          "inspection_recommended"
        ],
        "title": "InspectionTicketResponse",
        "description": "Enriched field inspection ticket work order."
      },
      "MeterReading": {
        "properties": {
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "Calendar date or ISO timestamp (e.g. '2016-10-30')"
          },
          "consumption_kwh": {
            "type": "number",
            "title": "Consumption Kwh",
            "description": "Measured daily electricity consumption in kilowatt-hours (kWh)"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "consumption_kwh"
        ],
        "title": "MeterReading",
        "description": "Individual daily smart meter consumption reading."
      },
...
      "SingleMeterPredictionRequest": {
        "properties": {
          "meter_id": {
            "type": "string",
            "minLength": 1,
            "title": "Meter Id",
            "description": "Unique electricity meter or account identifier"
          },
          "readings": {
            "items": {
              "$ref": "#/components/schemas/MeterReading"
            },
            "type": "array",
            "title": "Readings",
            "description": "Chronological series of daily meter consumption readings"
          },
          "evaluation_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Date",
            "description": "Target evaluation date (defaults to latest timestamp in readings)"
          },
          "tariff": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Tariff",
            "description": "Optional tariff rate override (currency/kWh)"
          },
          "customer_type": {
            "type": "string",
            "title": "Customer Type",
            "description": "Customer classification (e.g. residential, commercial, industrial)",
            "default": "standard"
          },
          "feeder_id": {
            "type": "string",
            "title": "Feeder Id",
            "description": "Distribution substation feeder identifier",
            "default": "unknown"
          },
          "dispatch_cost": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispatch Cost",
            "description": "Field inspection dispatch cost override"
          },
          "decision_rule": {
            "type": "string",
            "title": "Decision Rule",
            "description": "Decision policy: 'env', 'cost_threshold', or 'fixed_threshold'",
            "default": "env"
          },
          "fixed_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Fixed Threshold",
            "description": "Probability cutoff if fixed_threshold rule is active",
            "default": 0.5
          },
          "include_explanation": {
            "type": "boolean",
            "title": "Include Explanation",
            "description": "Whether to compute Tree-SHAP attributions and narratives",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "meter_id",
          "readings"
        ],
        "title": "SingleMeterPredictionRequest",
        "description": "Request payload for evaluating an individual smart meter time series."
      },
      "SingleMeterPredictionResponse": {
        "properties": {
          "meter_id": {
            "type": "string",
            "title": "Meter Id"
          },
          "evaluation_timestamp": {
            "type": "string",
            "title": "Evaluation Timestamp"
          },
          "model": {
            "$ref": "#/components/schemas/ModelMetadataResponse"
          },
          "data_quality": {
            "$ref": "#/components/schemas/DataQualitySummary"
          },
          "prediction": {
            "$ref": "#/components/schemas/PredictionOutput"
          },
          "financial": {
            "$ref": "#/components/schemas/FinancialOutput"
          },
          "decision": {
            "$ref": "#/components/schemas/DecisionOutput"
          },
          "explanation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExplanationOutput"
              },
              {
                "type": "null"
              }
            ]
          },
          "processing_time_ms": {
            "type": "number",
            "title": "Processing Time Ms",
            "description": "Total server processing latency in milliseconds"
          }
        },
        "type": "object",
        "required": [
          "meter_id",
          "evaluation_timestamp",
          "model",
          "data_quality",
          "prediction",
          "financial",
          "decision",
          "processing_time_ms"
        ],
        "title": "SingleMeterPredictionResponse",
        "description": "Full canonical response object for a single meter prediction request."
      },
...
    }
  }
}